Remove termination file
